fix: fix knip errors
