--- src/modules/listeners/model/listeners-factory.ts
@@ -1,35 +1,0 @@
-import { ScheduledListener } from "../classes/scheduled-listener";
-import { WaitingListener } from "../classes/waiting-listener";
-import { ListenersStore } from "./listeners-array";
-import { ListenerCallback } from "../types";
-import { abort } from "@modules/test-aborting";
-
-export class ListenersFactory<Params> {
-    public constructor(private readonly callback: ListenerCallback<Params>) {}
-
-    public wait(title: string, delay: number, params?: Params) {
-        const id = ListenersStore.id;
-
-        const listener = new WaitingListener(id, title, delay, () => this.callback({ params, abort }));
-
-        ListenersStore.add(listener);
-
-        return () => {
-            ListenersStore.remove(id);
-            listener.cancel();
-        };
-    }
-
-    public schedule(title: string, delay: number, params?: Params) {
-        const id = ListenersStore.id;
-
-        const listener = new ScheduledListener(id, title, delay, () => this.callback({ params, abort }));
-
-        ListenersStore.add(listener);
-
-        return () => {
-            ListenersStore.remove(id);
-            listener.cancel();
-        };
-    }
-}
